hw03 changed

+++ hw03/src/include/hw03.h
@@ -4,4 +4,4 @@
 
 void hw03(char* str, const char* pattern);
 int _get_start_index(char* str, const char* pattern);
-bool _is_equals(char* str, const char* pattern, int32_t i);
+bool _is_equals(char* str, const char* pattern, size_t i);
